Refactor database related files
